+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Version 2.1
 - Issue #139: Modify `GetDfsInfoInternal()` to use `SafeGlobalMemoryBufferHandle` instead of `IntPtr`.
 - Issue #141: Remove obsolete Resources (resx) string messages.
 - Issue #142: Move literal strings to Resources (resx)
+- Issue #144: Add `DirectoryInfo.EnumerateXxx()` methods with support for DirectoryEnumerationOptions enum.
 
 ### Bugs Fixed